# contracts/openapi/public-v1.yaml
openapi: 3.1.0
info:
  title: DengueSense Public API
  version: 1.0.0-draft.1
  summary: API công khai (qua gateway) cho dashboard DengueSense — phạm vi Đợt 1 (xem rủi ro).
  description: |
    **Nguồn sự thật của hợp đồng giữa `gateway` ↔ `dashboard`** (docs/09 §6.10).
    Code Go sinh từ file này (oapi-codegen), dashboard sinh type từ file này (openapi-typescript).

    Quy ước chung (docs/09 §6):
    - Trường JSON `snake_case`. Tháng dạng `"YYYY-MM"`, thời điểm RFC 3339 UTC.
    - Lỗi trả `application/problem+json` (RFC 9457) với `code` ổn định — danh mục ở `contracts/errors.md`.
    - Mọi response chứa số liệu mô hình có `meta` (provenance). Xác suất luôn đi cùng `base_rate`.
    - Client PHẢI bỏ qua trường lạ trong response (thêm trường tuỳ chọn không phải thay đổi phá vỡ).
    - Mọi `POST` tạo tài nguyên / có tác dụng phụ nhận header `Idempotency-Key`.

    Phạm vi bản này: đăng nhập, danh mục, quan sát, lượt dự báo, bản đồ rủi ro, giải thích, model card.
    Cảnh báo / hồ sơ / phân bổ / dự thảo / duyệt (Đợt 2–3) được thêm bằng PR hợp đồng riêng khi tới lượt.
servers:
  - url: /api/v1
    description: Cùng origin với dashboard (qua reverse proxy)
...
security:
  - bearerAuth: []
paths:
...
  /alerts:
    get:
      tags: [alerts]
      operationId: listAlerts
      summary: Danh sách cảnh báo dịch
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, all]
          description: Lọc theo trạng thái cảnh báo.
        - name: run_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Lọc theo lượt dự báo.
      responses:
        "200":
          description: Danh sách cảnh báo.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Alert" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /alerts/{alert_id}/confirm:
    post:
      tags: [alerts]
      operationId: confirmAlert
      summary: Xác nhận cảnh báo và tự động mở hồ sơ xử lý
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Xác nhận thành công.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /cases:
    get:
      tags: [cases]
      operationId: listCases
      summary: Danh sách hồ sơ xử lý dịch
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, completed, archived, all]
          description: Lọc theo trạng thái hồ sơ.
      responses:
        "200":
          description: Danh sách hồ sơ.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Case" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [cases]
      operationId: createCase
      summary: Tạo hồ sơ xử lý dịch mới
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCaseRequest" }
      responses:
        "201":
          description: Tạo hồ sơ thành công.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Case" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /cases/{case_id}:
    get:
      tags: [cases]
      operationId: getCase
      summary: Chi tiết hồ sơ xử lý dịch
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chi tiết hồ sơ.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaseDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /cases/{case_id}/allocation-plans:
    post:
      tags: [cases]
      operationId: createAllocationPlan
      summary: Tạo phương án phân bổ nguồn lực cho hồ sơ
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAllocationPlanRequest" }
      responses:
        "201":
          description: Tạo phương án thành công.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AllocationPlan" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /cases/{case_id}/drafts:
    post:
      tags: [drafts]
      operationId: createDraft
      summary: Tạo dự thảo văn bản cho hồ sơ (B2B hoặc B2G)
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDraftRequest" }
      responses:
        "201":
          description: Tạo dự thảo thành công.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDraft" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /cases/{case_id}/drafts/generate:
    post:
      tags: [drafts]
      operationId: generateDraftForCase
      summary: Dùng AI sinh dự thảo văn bản (B2B hoặc B2G) cho hồ sơ
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateDraftCaseRequest" }
      responses:
        "201":
          description: Sinh dự thảo thành công và lưu vào hồ sơ.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDraft" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /drafts/{draft_id}:
    get:
      tags: [drafts]
      operationId: getDraft
      summary: Chi tiết dự thảo văn bản
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chi tiết dự thảo.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDraft" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [drafts]
      operationId: updateDraft
      summary: Cập nhật nội dung dự thảo (tăng version, tính lại content_hash)
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDraftRequest" }
      responses:
        "200":
          description: Cập nhật thành công.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDraft" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /drafts/{draft_id}/reviews:
    post:
      tags: [drafts]
      operationId: submitReview
      summary: Phê duyệt / Từ chối / Yêu cầu sửa dự thảo (Áp dụng quy tắc 4 mắt)
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitReviewRequest" }
      responses:
        "200":
          description: Phê duyệt thành công.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /drafts/{draft_id}/dispatch:
    post:
      tags: [drafts]
      operationId: dispatchApprovedOrder
      summary: Ban hành lệnh điều phối sau khi dự thảo đã được duyệt
      parameters:
        - name: draft_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchOrderRequest" }
      responses:
        "201":
          description: Ban hành lệnh thành công.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchOrder" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (EdDSA, 15 phút) lấy từ `/auth/login`.
...
  responses:
    ValidationError:
      description: Đầu vào không hợp lệ (`code` = `common.validation_error`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Chưa xác thực hoặc token hết hạn.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Không đủ quyền.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Không tìm thấy.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details (docs/09 §6.3).
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Mã lỗi ổn định dạng `<service>.<snake_case>` — danh mục ở contracts/errors.md.
          examples: ["common.validation_error"]
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Alert:
      type: object
      required: [id, run_id, province_id, created_at, status]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        province_id: { type: string }
        created_at: { type: string, format: date-time }
        status: { type: string, enum: [open, closed] }
    Case:
      type: object
      required: [id, title, status, created_by, created_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { type: string, enum: [open, completed, archived] }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/Case"
        - type: object
          properties:
            alert_ids:
              type: array
              items: { type: string, format: uuid }
            latest_plan_id:
              type: string
              format: uuid
            latest_draft_id:
              type: string
              format: uuid
    CreateCaseRequest:
      type: object
      required: [title, created_by]
      properties:
        title: { type: string }
        alert_ids:
          type: array
          items: { type: string, format: uuid }
        created_by: { type: string }
    CreateAllocationPlanRequest:
      type: object
      required: [budget, items]
      properties:
        budget: { type: number, format: double }
        items:
          type: array
          items:
            type: object
            required: [province_id, cases_pred, cost]
            properties:
              province_id: { type: string }
              cases_pred: { type: number, format: double }
              cost: { type: number, format: double }
    AllocationPlan:
      type: object
      required: [id, case_id, budget, total_cost, estimated_cases_prevented, allocations, created_at]
      properties:
        id: { type: string, format: uuid }
        case_id: { type: string, format: uuid }
        budget: { type: number, format: double }
        total_cost: { type: number, format: double }
        estimated_cases_prevented: { type: number, format: double }
        allocations:
          type: array
          items:
            type: object
            required: [province_id, amount, explanation]
            properties:
              province_id: { type: string }
              amount: { type: number, format: double }
              explanation: { type: string }
        created_at: { type: string, format: date-time }
    DispatchDraft:
      type: object
      required: [id, case_id, version, draft_type, title, content, content_hash, status, created_by, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        case_id: { type: string, format: uuid }
        version: { type: integer }
        draft_type: { type: string, enum: [b2b, b2g] }
        title: { type: string }
        content: { type: string }
        content_hash: { type: string }
        status:
          type: string
          enum: [DRAFT, PENDING_REVIEW, CHANGES_REQUESTED, APPROVED, REJECTED]
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateDraftRequest:
      type: object
      required: [draft_type, title, content, created_by]
      properties:
        draft_type: { type: string, enum: [b2b, b2g] }
        title: { type: string }
        content: { type: string }
        created_by: { type: string }
    GenerateDraftCaseRequest:
      type: object
      required: [draft_type]
      properties:
        draft_type:
          type: string
          enum: [b2b, b2g]
        title:
          type: string
        extra_instructions:
          type: string
    UpdateDraftRequest:
      type: object
      required: [content, updated_by]
      properties:
        title: { type: string }
        content: { type: string }
        updated_by: { type: string }
    SubmitReviewRequest:
      type: object
      required: [reviewer_id, action, draft_version]
      properties:
        reviewer_id: { type: string }
        draft_version: { type: integer }
        action:
          type: string
          enum: [approve, reject, request_changes]
        note: { type: string }
    ReviewResult:
      type: object
      required: [id, draft_id, draft_version, reviewer_id, action, reviewed_at, draft_status]
      properties:
        id: { type: string, format: uuid }
        draft_id: { type: string, format: uuid }
        draft_version: { type: integer }
        reviewer_id: { type: string }
        action: { type: string, enum: [approve, reject, request_changes] }
        note: { type: string }
        reviewed_at: { type: string, format: date-time }
        draft_status: { type: string }
    DispatchOrderRequest:
      type: object
      required: [recipient, dispatched_by]
      properties:
        recipient: { type: string }
        dispatched_by: { type: string }
    DispatchOrder:
      type: object
      required: [id, case_id, draft_id, recipient, content, content_hash, status, created_at]
      properties:
        id: { type: string, format: uuid }
        case_id: { type: string, format: uuid }
        draft_id: { type: string, format: uuid }
        recipient: { type: string }
        content: { type: string }
        content_hash: { type: string }
        status: { type: string, enum: [PENDING, SENDING, SENT, DELIVERY_FAILED] }
        created_at: { type: string, format: date-time }